+++ api-builder-plugin-fn-redis/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Allows loading the plugin in development mode with bad Redis connection
 
 ## [0.0.3] 2020-04-24
 ### Added